Fixed numerous style problems.

+++ modules/services-ejb3/src/main/java/ccc/services/ejb3/ResourcesEJB.java
@@ -257,7 +257,7 @@ public class ResourcesEJB
     public void createWorkingCopy(final UUID resourceId,
                                   final ResourceDto pu)
     throws RestException {
-        createWorkingCopy(resourceId, pu.getRevision());
+        createWorkingCopy(resourceId, pu.getRevision().longValue());
     }
